Fix SendKeys ordering
- Created `SelectManyConcat` like `FlatMapConcat` in other languages
- Used `SelectManyConcat` to stop applying pit strategies concurrently
- Moved all extension methods in a common project

http://introtorx.com/Content/v1.0.10621.0/08_Transformation.html

> The difference with IObservable<T> sequences is that the call to
> the SelectMany's selector function is not blocking and the result
> sequence can produce values over time. This means that subsequent
> 'child' sequences can overlap.
