+++ RaceDirector.sln
@@ -47,6 +47,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PitCrew.Server.Tests", "tes
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PitCrew.Interface", "src\PitCrew.Interface\PitCrew.Interface.csproj", "{07F03868-92B8-452D-AD93-71F0E8B47281}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RaceDirector.Reactive", "src\RaceDirector.Reactive\RaceDirector.Reactive.csproj", "{B33AD34C-3A9A-4812-8AB7-82B479F5FD4D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RaceDirector.Reactive.Tests", "tests\RaceDirector.Reactive.Tests\RaceDirector.Reactive.Tests.csproj", "{75CF5E50-68D8-4155-A3FC-D9D9ECB845D1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -129,6 +133,14 @@ Global
 		{07F03868-92B8-452D-AD93-71F0E8B47281}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07F03868-92B8-452D-AD93-71F0E8B47281}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07F03868-92B8-452D-AD93-71F0E8B47281}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B33AD34C-3A9A-4812-8AB7-82B479F5FD4D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B33AD34C-3A9A-4812-8AB7-82B479F5FD4D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B33AD34C-3A9A-4812-8AB7-82B479F5FD4D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B33AD34C-3A9A-4812-8AB7-82B479F5FD4D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{75CF5E50-68D8-4155-A3FC-D9D9ECB845D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{75CF5E50-68D8-4155-A3FC-D9D9ECB845D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{75CF5E50-68D8-4155-A3FC-D9D9ECB845D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{75CF5E50-68D8-4155-A3FC-D9D9ECB845D1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
