+++ src/DeviceIO.Plugin/Pipeline/DeviceIoNode.cs
@@ -1,9 +1,8 @@
 ﻿using System.Reactive;
-using System.Reactive.Concurrency;
-using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using Microsoft.Extensions.Logging;
 using RaceDirector.Pipeline;
+using RaceDirector.Reactive;
 
 namespace RaceDirector.DeviceIO.Pipeline;
 
@@ -26,7 +25,7 @@ public class DeviceIoNode : INode
         _logger = logger;
         var subject = new Subject<GameAction>();
         subject
-            .ObserveOn(NewThreadScheduler.Default)
+            .SpacedBy(_configuration.WaitBetweenKeys)
             .Subscribe(Observer.Create<GameAction>(SendKeysToGame));
         GameActionObserver = subject;
     }
@@ -36,7 +35,6 @@ public class DeviceIoNode : INode
         if (_configuration.KeyMappings.TryGetValue(ga.ToString(), out var keys))
         {
             SendKeys.SendWait(keys);
-            Thread.Sleep(_configuration.WaitBetweenKeys);
             _logger.LogTrace("Received {GameAction} sent {Keys}", ga, keys);
         }
         else
